Make q10 transaction safe
When using a TableLoad directly in a radix join, the underlying table's
modifications are directly visible to the radix join's operators. This
might pose problems. Using a validatePositions operator in between
exposes the table as a pointer calculator the join and modifications of
the table are hidden to the join.