Refactored default use of jemalloc in settings
Just adding jemalloc in the toString method of Settings does not seem like a
good fit to me. I made the setting more obvious and explicit. It broke up the
general print the entire array simply to a string and required explicit
handling of USE_JE_MALLOC key. Worth it.