Improve server shutdown / graceful shutdown
No more killall for hyrise. Instead of waiting for random amounts of time,
explictly wait for a binary to finish. This is important, because even
killing a Hyrise process does not mean it is gone immediately.

Furthermore, users might get hung up in their fireQuery methods. Joining them
without a timeout might make a graceful termination impossible. We added a
timeout and terminate the user processes after the timeout.